move db_host variable to roles and add new variables
Having db_host set to members of the database group prevent
from emptying that inventory group so it cant' be seet in
group_vars/all. Instead using the defaults of roles sounds
a viable option. I also  prefered individual strong variables
instead of strutured data as it's easier to pass as a command
line argument.